requirements.txt
flask
flask-cors
crewai
langchain-groq
python-dotenv
edge-tts
litellm
playwright
supabase
moviepy
